First pass at a simple metric for variable objects
This is a very simple metric that simulates sinusoidal
varibility and tries to recover the period of the
variability using the peak value of a Lomb-Scargle
periodogram.  The data are sampled using the cadence
of the input OpSim run.

Since this is just a demonstration and is intentionally simple, there
are several limitations:
1. The periods explored in the periodogram
are windowed to be +-1 day of the possible periods.
2. No attempt is made to simulate the errors on the measurements.
3. This reports the guess at the period not the error in the period
relative to the input period.